pyproject.toml
[project]
name = 'lacosmic'
description = 'Laplacian Cosmic Ray Identification'
readme = 'README.rst'
license-files = ['LICENSE.rst']
authors = [
    {name = 'Larry Bradley', email = 'larry.bradley@gmail.com'},
]
keywords = [
    'astronomy',
    'astrophysics',
    'images',
    'cosmic rays',
    'cosmic-ray removal',
    'filtering',
]
classifiers = [
    'Intended Audience :: Science/Research',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Topic :: Scientific/Engineering :: Astronomy',
]
dynamic = ['version']
requires-python = '>=3.11'
dependencies = [
    'numpy >= 1.26',
    'astropy >= 5.3',
    'scipy >= 1.12.0',
]

[project.urls]
Homepage = 'https://github.com/larrybradley/lacosmic'
Documentation = 'https://lacosmic.readthedocs.io/en/stable/'

[project.optional-dependencies]
test = [
    'pytest-astropy >= 0.11',
    'tox > 4.0',
]
docs = [
    'lacosmic',
    'sphinx',
    'sphinx-astropy >= 1.9',
    'matplotlib >= 3.8',
]
dev = [
   'lacosmic[docs,test]',
   'pre-commit > 4.0',
]

[build-system]
requires = [
    'setuptools >= 77.0',
    'setuptools_scm >= 8.0',
    'numpy >= 2.0.0',
]
build-backend = 'setuptools.build_meta'

[tool.setuptools_scm]
write_to = 'lacosmic/version.py'

[tool.setuptools]
zip-safe = false
include-package-data = false

[tool.setuptools.packages.find]
namespaces = false

[tool.setuptools.package-data]
'lacosmic' = [
    'data/*',
]

[tool.pytest.ini_options]
minversion = 7.0
testpaths = [
    'lacosmic',
    'docs',
]
norecursedirs = [
    'docs/_build',
    'extern',
]
astropy_header = true
doctest_plus = 'enabled'
text_file_format = 'rst'
addopts = [
    '-ra',
    '--color=yes',
    '--doctest-rst',
    '--strict-config',
    '--strict-markers',
]
log_cli_level = 'INFO'
xfail_strict = true
remote_data_strict = true
filterwarnings = [
    'error',  # turn warnings into exceptions
    'ignore:numpy.ndarray size changed:RuntimeWarning',
]

[tool.coverage.run]
omit = [
    'lacosmic/_astropy_init*',
    'lacosmic/conftest.py',
    'lacosmic/*setup_package*',
    'lacosmic/tests/*',
    'lacosmic/*/tests/*',
    'lacosmic/extern/*',
    'lacosmic/version*',
    '*/lacosmic/_astropy_init*',
    '*/lacosmic/conftest.py',
    '*/lacosmic/*setup_package*',
    '*/lacosmic/tests/*',
    '*/lacosmic/*/tests/*',
    '*/lacosmic/extern/*',
    '*/lacosmic/version*',
]

[tool.coverage.report]
exclude_lines = [
    'pragma: no cover',
    'except ImportError',
    'raise AssertionError',
    'raise NotImplementedError',
    'def main\\(.*\\):',
    'pragma: py{ignore_python_version}',
    'def _ipython_key_completions_',
]

[tool.repo-review]
ignore = [
    'MY',  # ignore MyPy
    'PC110',  # ignore using black or ruff-format in pre-commit
    'PC111',  # ignore using blacken-docs in pre-commit
    'PC140',  # ignore using mypy in pre-commit
    'PC180',  # ignore using prettier in pre-commit
    'PC901',  # ignore using custom pre-commit update message
    'PY005',  # ignore having a tests/ folder
]

[tool.isort]
skip_glob = [
    'lacosmic/*__init__.py*',
]
known_first_party = [
    'lacosmic',
    'extension_helpers',
]
use_parentheses = true

[tool.black]
force-exclude = """
(
  .*
)
"""

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.numpydoc_validation]
checks = [
    'all',   # report on all checks, except the below
    'ES01',  # missing extended summary
    'EX01',  # missing "Examples"
    'SA01',  # missing "See Also"
    'SA04',  # missing "See Also" description
    'RT02',  # only type in "Returns" section (no name)
    'SS06',  # single-line summary
    'RT01',  # do not require return type for lazy properties
]

# don't report on objects that match any of these regex;
# remember to use single quotes for regex in TOML
exclude = [
    '__init__',
    '\._.*',  # private functions/methods
    '^test_*',  # test code
    '^conftest.*$',  # pytest configuration
]

[tool.docformatter]
    wrap-summaries = 72
    pre-summary-newline = true
    make-summary-multi-line = true

[tool.ruff]
line-length = 79

[tool.ruff.lint.pylint]
max-statements = 130

[tool.ruff.lint]
select = ['ALL']
ignore = [
    'ANN',
    'D200',
    'D205',
    'FBT002',
    'PLC0415',  # import-outside-top-level
    'PLR0913',
    'PTH',
    'Q000',
]

[tool.ruff.lint.per-file-ignores]
'__init__.py' = ['D104', 'I']
'conftest.py' = ['D103', 'I']
'docs/conf.py' = ['ERA001', 'INP001', 'TRY400']
'test_*.py' = ['D']

[tool.ruff.lint.pydocstyle]
convention = 'numpy'
